Add a gyro mode switch.
Depending on the mode, return the gyro values either as momentary
angular velocity [deg/s], or as accumulated rotation [deg]. The latter
is the default now. This applies to local, world, and player space, the
gravity vector readings are unaffected.

Also added a 'reset' message to reset the accumulated rotation to the
origin, and renamed the calibration 'reset' message to 'clear'.